fixed api.Fetch cache values assign error and small internal renaming
