--- internal/command/callbacks/responses.go
@@ -1,11 +1,0 @@
-package callbacks
-
-type MapResponse struct {
-	Count    int    `json:"count"`
-	Next     string `json:"next"`
-	Previous string `json:"previous"`
-	Results  []struct {
-		Name string `json:"name"`
-		URL  string `json:"url"`
-	} `json:"results"`
-}
